[220914] BE: Respository 타입값 수정

See merge request s07-bigdata-recom-sub2/S07P22B203!15

+++ BE/hp-domain/src/main/java/com/ssafy/hp/auth/AuthRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.hp.auth;
 import com.ssafy.hp.auth.domain.Auth;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthRepository extends CrudRepository<Auth, Long> {
+public interface AuthRepository extends CrudRepository<Auth, Integer> {
 }
