+++ BE/hp-domain/src/main/java/com/ssafy/hp/user/UserExerciseRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.hp.user;
 import com.ssafy.hp.user.domain.UserExercise;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserExerciseRepository extends JpaRepository<UserExercise, Long> {
+public interface UserExerciseRepository extends JpaRepository<UserExercise, Integer> {
 }
